Tests/LibWeb/Text/expected/css/pad-descriptor-calculated-value.txt

@counter-style foo { pad: calc(1 + sign(1em - 1px)) "0"; }
